Added scaffolding for a MediaBrowser
Added ability to play a song by clicking the title, and the ability to pause or play the current song.

Added a custom Application.
Added a MediaBrowserService.
ActivityMain now interacts with a MediaBrowserService.
Added FragmentControls to control playback.
FragmentMusicList can play a song when the user clicks one.
Added a MediaPlayerQueue to hold a queue of MediaPlayers.
MediaPlayerRepository now has play and pause capabilities and can release the enclosed MediaPlayer.

Tests are probably broken and need to be run.
New tests need to be made for the new functionality.