Version 0.3.3 implemented and tested
Songs now shuffle via a probability distribution.
Songs can now be skipped.
When a song is skipped, the probability that it plays again is reduced and the amount the probability is reduced can be changed by the user.
A settings for respecting audio focus of other apps was added.
The search text when searching music is now saved across lifecycle restarts.
A dialog with a description of why permission is needed and a link to the settings replaced the toast with a description of why permission is needed when permission is denied.
The app no longer crashes when there is no music on the device.

Added a flavor of the app that provides no music to the app.
Switched from kapt to ksp.
Switched to Material buttons.
The probability map can now scale probability by any fraction.
All buttons are no longer all caps.